requirements.txt
aiofiles==0.4.0
CacheControl==0.12.6
cachy==0.3.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
cleo==0.7.6
clikit==0.4.1
cryptography==2.8
gunicorn==20.0.4
h11==0.8.1
h2==3.1.1
hpack==3.0.0
html5lib==1.0.1
httpcore==0.3.0
httptools==0.0.13
hyperframe==5.2.0
idna==2.8
jeepney==0.4.2
jsonschema==3.2.0
keyring==20.0.1
lockfile==0.12.2
msgpack==0.6.2
multidict==4.6.1
pastel==0.1.1
pexpect==4.8.0
pkginfo==1.5.0.1
poetry==1.0.2
ptyprocess==0.6.0
pycparser==2.19
pylev==1.3.0
pyparsing==2.4.6
pyrsistent==0.14.11
python-decouple==3.3
python-dotenv==0.10.3
redis==3.3.11
requests==2.22.0
requests-async==0.5.0
requests-toolbelt==0.8.0
rfc3986==1.3.2
sanic==19.9.0
SecretStorage==3.1.2
shellingham==1.3.1
tomlkit==0.5.8
ujson==1.35
urllib3==1.25.7
uvloop==0.14.0
webencodings==0.5.1
websockets==8.1
